+++ client/shared/src/actions/ActionItem.test.tsx
@@ -16,6 +16,7 @@ describe('ActionItem', () => {
     test('non-actionItem variant', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 location={history.location}
@@ -29,6 +30,7 @@ describe('ActionItem', () => {
     test('actionItem variant', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -43,6 +45,7 @@ describe('ActionItem', () => {
     test('noop command', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 location={history.location}
@@ -56,6 +59,7 @@ describe('ActionItem', () => {
     test('pressed toggle actionItem', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'a', command: 'c', actionItem: { pressed: true, label: 'b' } }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -70,6 +74,7 @@ describe('ActionItem', () => {
     test('non-pressed actionItem', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'a', command: 'c', actionItem: { pressed: false, label: 'b' } }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -84,6 +89,7 @@ describe('ActionItem', () => {
     test('title element', () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -101,6 +107,7 @@ describe('ActionItem', () => {
 
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -130,6 +137,7 @@ describe('ActionItem', () => {
 
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -157,6 +165,7 @@ describe('ActionItem', () => {
     test('run command with error', async () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -182,6 +191,7 @@ describe('ActionItem', () => {
     test('run command with error with showInlineError', async () => {
         const component = renderer.create(
             <ActionItem
+                active={true}
                 action={{ id: 'c', command: 'c', title: 't', description: 'd', iconURL: 'u', category: 'g' }}
                 telemetryService={NOOP_TELEMETRY_SERVICE}
                 variant="actionItem"
@@ -210,6 +220,7 @@ describe('ActionItem', () => {
 
             const component = renderer.create(
                 <ActionItem
+                    active={true}
                     action={{ id: 'c', command: 'open', commandArguments: ['https://example.com/bar'], title: 't' }}
                     telemetryService={NOOP_TELEMETRY_SERVICE}
                     location={history.location}
@@ -225,6 +236,7 @@ describe('ActionItem', () => {
 
             const component = renderer.create(
                 <ActionItem
+                    active={true}
                     action={{ id: 'c', command: 'open', commandArguments: ['https://other.com/foo'], title: 't' }}
                     telemetryService={NOOP_TELEMETRY_SERVICE}
                     location={history.location}
@@ -240,6 +252,7 @@ describe('ActionItem', () => {
 
             const component = renderer.create(
                 <ActionItem
+                    active={true}
                     action={{ id: 'c1', command: 'whatever', title: 'primary' }}
                     altAction={{ id: 'c2', command: 'open', commandArguments: ['https://other.com/foo'], title: 'alt' }}
                     telemetryService={NOOP_TELEMETRY_SERVICE}