+++ client/web/src/enterprise/batches/list/BatchChangesListEmpty.tsx
@@ -126,7 +126,7 @@ export const BatchChangesListEmpty: React.FunctionComponent<BatchChangesListEmpt
                 className="percy-hide chromatic-ignore"
                 width="100%"
                 height="600"
-                src="https://www.youtube.com/embed/EfKwKFzOs3E"
+                src="https://www.youtube-nocookie.com/embed/eOmiyXIWTCw"
                 frameBorder="0"
                 allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
                 allowFullScreen={true}