+++ cmd/gitserver/server/server.go
@@ -1006,7 +1006,7 @@ func (s *Server) handleP4Exec(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Make sure credentials are valid before heavier operation
-	err := p4pingWithLogin(r.Context(), req.P4Port, req.P4User, req.P4Passwd)
+	err := p4pingWithTrust(r.Context(), req.P4Port, req.P4User, req.P4Passwd)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
@@ -1122,6 +1122,7 @@ func (s *Server) p4exec(w http.ResponseWriter, r *http.Request, req *protocol.P4
 	cmd.Env = append(os.Environ(),
 		"P4PORT="+req.P4Port,
 		"P4USER="+req.P4User,
+		"P4PASSWD="+req.P4Passwd,
 	)
 	cmd.Stdout = stdoutW
 	cmd.Stderr = stderrW
@@ -1486,7 +1487,7 @@ func newURLRedactor(parsedURL *url.URL) *urlRedactor {
 // Sensitive strings are replaced with "<redacted>".
 func (r *urlRedactor) redact(message string) string {
 	for _, s := range r.sensitive {
-		message = strings.Replace(message, s, "<redacted>", -1)
+		message = strings.ReplaceAll(message, s, "<redacted>")
 	}
 	return message
 }