+++ internal/encryption/keyring/ring.go
@@ -17,6 +17,9 @@ var (
 	defaultRing Ring
 )
 
+// Default returns the default keyring, if you can avoid using this from arbitrary points in your code, please do!
+// we would rather inject the individual keys as dependencies when initialised from main(), but if that's impractical
+// it's fine to use this.
 func Default() Ring {
 	mu.RLock()
 	defer mu.RUnlock()